requirements.txt
# Project dependencies for acs-lab
python-dotenv
msal
requests
azure-communication-email
azure-communication-sms
